+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.1.0 (2014-09-17)
+## 0.2.0 (2014-??-??)
+
+ - added support for JSON.stringify's `replacer` parameter
+ - improved error message
+
+## 0.1.0 (2014-09-18)
 
 Initial release